+++ include/crpropa/Grid.h
@@ -119,7 +119,7 @@ public:
 
 	/** set the type of interpolation between grid points.
 	 * @param i: interpolationType (TRILINEAR, TRICUBIC, NEAREST_NEIGHBOUR) */
-	void setInterpolationType(interpolationType i){
+	void setInterpolationType(interpolationType i) {
 		ipol = i;
 	}
 };
@@ -140,6 +140,7 @@ class Grid: public Referenced {
 	Vector3d origin; /**< Origin of the volume that is represented by the grid. */
 	Vector3d gridOrigin; /**< Grid origin */
 	Vector3d spacing; /**< Distance between grid points, determines the extension of the grid */
+	bool clipVolume; /**< If set to true, all values outside of the grid will be 0*/
 	bool reflective; /**< If set to true, the grid is repeated reflectively instead of periodically */
 	interpolationType ipolType; /**< Type of interpolation between the grid points */
 
@@ -154,6 +155,7 @@ public:
 		setGridSize(N, N, N);
 		setSpacing(Vector3d(spacing));
 		setReflective(false);
+		setClipVolume(false);
 	}
 
 	/** Constructor for non-cubic grid
@@ -168,6 +170,7 @@ public:
 		setGridSize(Nx, Ny, Nz);
 		setSpacing(Vector3d(spacing));
 		setReflective(false);
+		setClipVolume(false);
 	}
 
 	/** Constructor for non-cubic grid with spacing vector
@@ -182,6 +185,7 @@ public:
 		setGridSize(Nx, Ny, Nz);
 		setSpacing(spacing);
 		setReflective(false);
+		setClipVolume(false);
 	}
 
 	/** Constructor for GridProperties
@@ -215,12 +219,17 @@ public:
 		reflective = b;
 	}
 
+	// If set to true, all values outside of the grid will be 0.
+	void setClipVolume(bool b) {
+		clipVolume = b;
+	}
+
 	/** Change the interpolation type to the routine specified by the user. Check if this routine is
 		contained in the enum interpolationType and thus supported by CRPropa.*/
 	void setInterpolationType(interpolationType ipolType) {
 		if (ipolType == TRILINEAR || ipolType == TRICUBIC || ipolType == NEAREST_NEIGHBOUR) {
 			this->ipolType = ipolType;
-			if ((ipolType == TRICUBIC) && (std::is_same<T, Vector3d>::value)){
+			if ((ipolType == TRICUBIC) && (std::is_same<T, Vector3d>::value)) {
 				KISS_LOG_WARNING << "Tricubic interpolation on Grid3d works only with float-precision, doubles will be downcasted";
 		}
 		} else {
@@ -262,12 +271,22 @@ public:
 	  By default this it the trilinear interpolation. The user can change the
 	  routine with the setInterpolationType function.*/
 	T interpolate(const Vector3d &position) {
-	if (ipolType == TRICUBIC)
-		return tricubicInterpolate(T(), position);
-	else if (ipolType == NEAREST_NEIGHBOUR)
-		return closestValue(position);
-	else
-		return trilinearInterpolate(position);
+		// check for volume
+		if (clipVolume) {
+			Vector3d edge = origin + Vector3d(Nx, Ny, Nz) * spacing;
+			bool isInVolume = (position.x >= origin.x) && (position.x <= edge.x);
+			isInVolume &= (position.y >= origin.y) && (position.y <= edge.y);
+			isInVolume &= (position.z >= origin.z) && (position.z <= edge.z);
+			if (!isInVolume) 
+				return T(0.);
+		} 
+
+		if (ipolType == TRICUBIC)
+			return tricubicInterpolate(T(), position);
+		else if (ipolType == NEAREST_NEIGHBOUR)
+			return closestValue(position);
+		else
+			return trilinearInterpolate(position);
 	}
 
 	/** Inspector & Mutator */