+++ src/massDistribution/ConstantDensity.cpp
@@ -26,11 +26,11 @@ double ConstantDensity::getDensity(const Vector3d &position) const {
 	if(isH2)
 		n += H2densitynumber;
 
-	// check if all densities are active and raise warning if not
-	if((isHI & isHII & isH2) == false){
+	// check if all densities are deactivated and raise warning if so
+	if((isHI || isHII || isH2) == false){
 		KISS_LOG_WARNING
-			<< "\nCalled getDensity on (partly) deactivated ConstantDensity \n"
-			<< "gas density model. Make sure this was intentional.";
+			<< "\nCalled getNucleonDensity on fully deactivated ConstantDensity "
+			<< "gas density model. In this case the density is allways set to 0. \n";
 	}
 
 	return n;
@@ -46,11 +46,11 @@ double ConstantDensity::getNucleonDensity(const Vector3d &position) const {
 	if(isH2)
 		n += 2*H2densitynumber;
 
-	// check if all densities are active and raise warning if not
-	if((isHI & isHII & isH2) == false){
+	// check if all densities are deactivated and raise warning if so
+	if((isHI || isHII || isH2) == false){
 		KISS_LOG_WARNING
-			<< "\nCalled getNucleonDensity on (partly) deactivated ConstantDensity \n"
-			<< "gas density model. Make sure this was intentional.";
+			<< "\nCalled getNucleonDensity on fully deactivated ConstantDensity "
+			<< "gas density model. In this case the density is allways set to 0. \n";
 	}
 	return n;
 }