+++ CHANGELOG.md
@@ -5,6 +5,9 @@
 
 ### New features:
 * new candidate property tagOrigin to trace back which source or which interaction created the candidate
+* new interace for massdistributions given on a Grid1f
+* grids can be restricted to the volume without repetition
+* sourceFeature to sample the source position from a given massdistribution
 
 ### Interface changes:
 * Weight column in hdf-Output is now called "W", which is the same as for TextOutput.